Add automatic repo sync during startup
It turns out that Heroku lets any web process run for 24 hours from the
time it last started. That means that the external sync_all script
running in a scheduler can easily get very far out of sync with the time
that Heroku restarts the web process. :/ To resolve this completely, I
added a new option to Homu: sync on startup. It runs the exact same code
as the external `sync_all` script does, but automatically and
immediately when the web process boots.

This isn't completely ideal, since it means that homu-on-heroku moves
off of the mainline Servo fork of Homu, but it has the positive effect
of making homu-on-heroku exactly as useful as homu.io is. This change
means that even on Heroku, Homu always knows about every single PR, and
never needs any manual or external syncing. We're using this change to
great effect in @bundlerbot with the Bundler and RubyGems open source
projects.